Add a sensitivity zone around best route in routage (experimental)
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@1175 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/vlmLine.h
@@ -42,7 +42,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 class vlmLine : public QGraphicsWidget
 { Q_OBJECT
     public:
-        vlmLine(Projection * proj, QGraphicsScene * myScene,int z_level);
+        vlmLine(Projection * proj, QGraphicsScene * myScene,double z_level);
 
         void addPoint(double lat,double lon);
         void addVlmPoint(vlmPoint point);
@@ -133,7 +133,7 @@ class vlmLine : public QGraphicsWidget
         GshhsReader *map;
         double lineWidth;
         QPainterPath myPath;
-        int myZvalue;
+        double myZvalue;
 };
